+++ src/ble/att_server.h
@@ -91,7 +91,7 @@ uint16_t att_server_get_mtu(hci_con_handle_t con_handle);
  * @note callback might happend during call to this function
  * @param callback_registration to point to callback function and context information
  * @param con_handle
- * @return 0 if ok, error otherwise
+ * @return ERROR_CODE_SUCCESS if ok, ERROR_CODE_UNKNOWN_CONNECTION_IDENTIFIER if handle unknown, and ERROR_CODE_COMMAND_DISALLOWED if callback already registered
  */
 int att_server_request_to_send_notification(btstack_context_callback_registration_t * callback_registration, hci_con_handle_t con_handle);
 
@@ -100,7 +100,7 @@ int att_server_request_to_send_notification(btstack_context_callback_registratio
  * @note callback might happend during call to this function
  * @param callback_registration to point to callback function and context information
  * @param con_handle
- * @return 0 if ok, error otherwise
+ * @return ERROR_CODE_SUCCESS if ok, ERROR_CODE_UNKNOWN_CONNECTION_IDENTIFIER if handle unknown, and ERROR_CODE_COMMAND_DISALLOWED if callback already registered
  */
 int att_server_request_to_send_indication(btstack_context_callback_registration_t * callback_registration, hci_con_handle_t con_handle);