+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - GAP: support reading RSSI for Classic+LE using gap_read_rssi. Emits GAP_EVENT_RSSI_MEASUREMENT
 - GAP: support setting Link Supervision Timetout for outgoing Classic connections using gap_set_link_supervision_timeout
 - Linked List: return if item was added for `btstack_linked_list_add[_tail]`
+- ATT Server: return error if request to send now for Notification or Indication already registered
 
 ## Changed
 - Bluetooth and BTstack Error Codes and Events: collect status codes in bluetooth.h and events in btstack_defines.h